rename GamePoint.py to game_points.py
